requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==4.2.12
django-debug-toolbar==4.4.6
idna==3.7
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
pytest==8.3.2
pytest-django==4.8.0
requests==2.32.3
sqlparse==0.5.1
urllib3==2.2.2
